Remove bbox in preference to bounds
Intead of having objects have a property named bbox which is an alias to
bounds.bbox they now just have a bounds property.  Objects that did not
have the alias e.g. Feature removed the bbox property in preference to
bounds.

BoundingBox#isOnBoundary was renamed to #isPositionOnBoundary to make
the method name less ambiguous.