Provide converter from GeoJson to Montage Geo
The converter GeoJsonToGeometryConverter will transform an object in
GeoJson notation to its associated value in Montage Geo.  The converter
can also revert a Montage Geo object to GeoJson notation.

The converter has a projection property that when set will be used to
transform the coordinates from the specified system to WGS 84 during the
conversation.  And from WGS 84 to the defined projection when reverted.

Initializers for Geometry objects now support supplying a projection.