Add original values to criteria parameters
The criteria created as part of the renderer now includes in its param-
eters the original values of the symbol, e.g., the classMaxValue (and
the classMinValue) of the class breaks renderer and the value of the
unique value renderer to match.

Unique Value Renderer also adds methods for accessing each of its field
properties.